feat: theme switching
A theme switching button is now added to the quick settings button grid.

An initial (WIP) light theme is added as an embedded gresource named
"way-shell-light.css".

When the theme is switched we run a script at
$HOME/.config/way-shell/on_theme_changed.sh with one argument which is
either "light" or "dark" to express the current theme.

The user can run any additional logic within this script when the theme
is switched.

Signed-off-by: ldelossa <louis.delos@gmail.com>