scripting: add on_workspace_new.sh script
When Way-Shell detects that a new window manager Workspace is created if
a script 'XDG_CONFIG_DIR/way-shell/on_workspace_new' is invoked with a
single argument, the workspace name.

This is useful when you'd like to use the workspace name to 'setup' the
workspace.

For instance you can write a script which looks for a workspace name and
then creates a particular window layout for this workspace.

Signed-off-by: ldelossa <louis.delos@gmail.com>