workspace-switcher: allow forced creation of workspace
The workspace-switcher is fuzzy-ish.

That means if you have a workspace named "testing-app-1" and you wanted
to create a new workspace "testing", there'd always be a selection
preset in the workspace-switcher (testing-app-1) and you could not
create "testing".

Now, if you want to create a workspace which name overlaps with an
existing one you can use 'Ctrl+Return' to ignore the search results and
create the workspace.

Signed-off-by: ldelossa <louis.delos@gmail.com>