refactor: reset message tray and quick settings animations
Prior to this comit both these components would no-op if a reveal
animation was in-flight.

This proved to be a bit problematic as the AdwAnimation state was
getting stuck in 'playing' despite the component was not visible.

Instead of no-op'ing we reset the animation state, ensuring we don't got
stuck in any particular state.

Signed-off-by: ldelossa <louis.delos@gmail.com>