refactor: various fixes for reliability
This commit fixes a few issues with widget initialization and
reinitialization.

Due to the previous removal of the QuickSettings and MessageTray
mediator structures, dependent widgets created during the construction
of these components, would receive NULL pointers for the respective
global getters.

Now, we set the global variable during construction, both QS and MT
structures are signletons, so this is OK.

Other then this, various cleanup around calling reinitialization methods
for no reason is performed.

Signed-off-by: ldelossa <louis.delos@gmail.com>