osd: animate OSD with AdwAnimation
To support SwayFX's shadows on gtk4-layer-shell surfaces we move our
GtkRevealers to AdwAnimation to drive OSD animations.

There seems to be a bit of artificating issue with SwayFX when a window
is removed, so lets switch to a 'pop up' animation for OSD which avoids
any of these artifacts issues.

Signed-off-by: ldelossa <louis.delos@gmail.com>