make: fix double gresources.o include
Prior to this commit running make two times in a row would cause the
gresources.o file to be included twice.

This is because the Makefile finds all .o files in /src, but we just
updated the Makefile to also hardcode src/gresources.o.

Instead, compile gresources.o outside of /src and update makefile/import
paths for related headers accordingly.

Signed-off-by: ldelossa <louis.delos@gmail.com>