refactor: rework app-switcher
The original app-switcher code was needlessly complex and a tiny bit
buggy.

This commit refactors the app-switcher code to simply bump the last
accessed app to the front of the box, removes tracking both a data model
array and a widget container, and finds a new way to correctly determine
which app should be selected when the app-switcher opens.