fix: audio fixes
This commit fixes several audio issues discovered in #23.

Way-Shell didn't understand that the wireplumber provided volume in a
cubic scale when a GET to the volume API was made.

We do the appropriate conversion to and from the linear and cubic scales
now.

Additionally, we had some event loops with volume scales. Adjusting
scales could cause an situation where the adjustment now wrote to the
mixer API, and a loop would occur. This is handled by blocking signals.

Signed-off-by: ldelossa <louis.delos@gmail.com>