refactor: move away from event controllers as a loop medigation measure
Previous to this commit we'd use an event controller to block signals
loops induced by adjusting sliders.

These event controllers proved to be a bit unreliable and required the
mouse to be tracked effectively.

Its better to just block the signals directly when the loop causing side
effect is called.

Signed-off-by: ldelossa <louis.delos@gmail.com>