fix use of logind's ListSessions
from the man page:
ListSessions() returns an array of all current sessions. The structures
in the array consist of the following fields: session id, user id, user
name, seat id, and session object path.
If a session does not have a seat attached, the seat id field will be an
empty string.