quick-settings,nm: add VPN toggling via quick settings
This commit adds a new VPN quick settings button and introduces the
necessary modifications to NetworkManagerService to support it.

The button will not be spawned if no VPN connections exist in
NetworkManager's database.

Once a VPN exists the VPN button will be spawned in the quick settings
area with a drop down allowing particular VPN connections to be toggled
on.

The VPN connection is *always* activated on the PrimaryDevice. This is
typically the device your machine uses to get to the internet.