activities,fix: map keys for next/prev search
Before this commit using tab in the activites view after some search
text was entered confused the UI.

Tab would move to the next element, but we want Tab to actually trigger
a move of the filtered search items.

Wire up a key controller to intercept Tab and make it do what we want.
Throw in ctrl-p/n while we are at it.

Signed-off-by: ldelossa <louis.delos@gmail.com>