sni: handle sni updates appropriately
The StatusNotifierItem DBus signals seem to be detached from actually
updating the cooresponding proxy's values.

It therefore became necessary to init the StatusNotifierItem with values
gained from the proxy attributes, but then perform a GetAll properties
call on the proxy when we receive a signal from the StatusNotifierItem
interface.

This commit handles this two-truths scenario, where we first init from
cached proxy attributes, then query DBus on signals

Signed-off-by: ldelossa <louis.delos@gmail.com>